Reorder the function calls in the 'then' clause of addContent in the renderPage function, for comprehension and performance considerations.
Sections links should be updated if the section is displayed. Content links should always be updated. Given the asynchronous nature of the updateLinks and updateEmbeddedLinks functions, they should be called as early as possible and embedded links may take the longest time to update considering potential recursive calls to updateEmbeddedLinks.